chapter11/code/bigquery/requirements.txt
google-cloud-bigquery
pandas
google-cloud-bigquery[pandas]
